Don't duplicate frozen parameters during predict() (#20851)
On the Jax backend we were not using donate_argnums during predict.
This works when a model is mostly trainable, but when a model is mostly
or all frozen, this will result in 2x the memory jump (which is why
we use donate_argnums for fit and evaluate).

This change adds donate_argnums to the predict function to avoid the
memory spike. But because this means all incoming state (including the
trainable variables) will be deleted by jax, this means we need to
sync the trainable variables state much like in fit and evaluate. An
alternative would be to change the predict_step signature (so we
could only donate non-trainable variables), but this would be a
breaking change and confusing.