Fix `BinaryAccuracy` to handle boolean inputs.
This is a follow up to https://github.com/keras-team/keras/pull/20782 and a replacement for https://github.com/keras-team/keras/pull/20782

We cannot cast `y_pred` and `y_true` to the expected output dtype in `MeanMetricWrapper`. Some metrics expect integers (indices or IDs for instance) and fail if `y_pred` and `y_true` are provided as floats.

It is the responsibility of the metric function to cast as needed.

In this case, the correct approach in `BinaryAccuracy` is to use the regular type promotion rules to ensure that the comparison between `y_pred` and `threshold` is done without losing precision. `ops.greater` already does the type promotion correctly. Previously, `threshold` was incorrectly cast to the `y_pred` dtype, which in this case would lower its precision.