Refactor ModelCheckpoint Save Logic (#21100)
The _save_model method combined the logic to determine if the checkpoint should be saved, and the logic to create the paths and save the checkpoint.

This commit separates the check the determine whether the checkpoint should be saved from the I/O logic, and in doing so resolves two bugs in the current implementation:

1) Host directory is created for every for save iteration, regardless of whether the model will be saved or not. For example, when `save_freq == 'epoch'` and `save_best_only == True`, a folder is created for every epoch, even though the model is only saved when the monitored condition is satisfied. This results in a large number of empty folders and makes it difficult to identify the most recently saved checkpoint.

With this commit, the directory to save the model or model weights is only created when necessary.

2) If save_best_only=True, and the monitored value is an np.ndarray or backend tensor, then it falls back to `save_best_only=False` and saves the model. However, in this scenario, it save saves the whole model without regard to the value of `self.save_weights_only`

This commit uses consistent save logic that always checks the value of `self.save_weights_only`.