`PyDataset` now implements `__iter__`.
It can now be iterated without failure when it has a finite number of batches.

Previously, it was iterable by accident due to a legacy feature of Python that predates the introduction of `__iter__`. Specifically, Python would see `__getitem__` implemented and would iterate by passing sequential indices. However, `__getitem__` was expected to throw an `IndexError` to indicate the end of the iterator as `__len__` is ignored by Python.

Instead, we return an iterator that knows about the length of the dataset.

Fixes https://github.com/keras-team/keras/issues/21151