Support non-tf.data.Dataset instances in multi-host setting.
It is assumed that the data is already pre-sharded across hosts.

If `distribution.auto_shard_dataset` is `True` but the input
dataset isn't a tf.data.Dataset, will error out as before.
Setting `auto_shard_dataset` to `False` will allow any
dataset type to be used.

Unfortunately we can't add a test for this in OSS since we don't
have a multi-process test fixture.  However, this was tested
internally.