Improve deserialization error about `@register_keras_serializable`.
There is a confusing case when a user has decorated their class or function with `@keras.saving.register_keras_serializable()`, but gets the error message: "Make sure custom classes are decorated with `@keras.saving.register_keras_serializable()`". This happens if the decorated function or class has not been imported/loaded and the decorator has not been run.

This augments the error message to explain this case.