Fix numpy tests for window ops, `angle`, `identity`, `eye`. (#21596)
JAX is introducing a change whereby some ops, in particular window ops, return a result of the default dtype for the current configuration (in particular based on the `enable_x64` flag).

For cross-backend consistency, we want the window ops to return `floatx` regardless of how JAX is configured.

Additionally, the unit tests need to be changed to not use the JAX returned dtype as the expected dtype.