Make `vectorized_map` op serializable. (#21597)
All ops that take tensors as inputs are in scope for serialization. They get serialized if they appear as part of a functional model.

The `VectorizedMap` op class needs a custom `get_config()` and `from_config()` to handle the `function` parameter, which is not a simple type. The fallback `get_config` logic only handles simple types.

Note that some other ops will need a similar change, which will come in a separate PR because it requires a another change.