Store optional arguments in auto-config
The current auto-config implementation only stores the supplied
arguments, but not the values of the optional arguments of the
constructor. As a consequence, deserializing a model fails when the
default value of an optional argument changes, even when the structure
of the model remains identical. This is problematic when we want to
provide better results for newly trained models, but want to keep old
models functional.

This commit extends the auto-config functionality to also store default
values, while otherwise being compatible with the previous
implementation.

Additional context:

In the BayesFlow library, we provide models for downstream use, and aim
to avoid changes that break loading of models. We rely on the
auto-config functionality to avoid the somewhat tedious and error-prone
hand-written specification of the configuration. To be able to change
defaults without breaking model loading, it would be great if defaults
were serialized in the auto-config as well.

See also https://github.com/bayesflow-org/bayesflow/issues/566.