auto-config: riase for variadic arguments, add tests
Variadic arguments (i.e., *args) and positional only arguments cannot be
passed via `cls(**config)`, so no automatic config can be created for
them without adding additional logic to `from_config`. This commit adds
an appropriate error message and tests for this case.

It also extends the test to contain an optional argument with a default
value. This allows testing the behavior defined in the previous commits.