Store defaults of optional arguments in auto-config (#21615)
* Store optional arguments in auto-config

The current auto-config implementation only stores the supplied
arguments, but not the values of the optional arguments of the
constructor. As a consequence, deserializing a model fails when the
default value of an optional argument changes, even when the structure
of the model remains identical. This is problematic when we want to
provide better results for newly trained models, but want to keep old
models functional.

This commit extends the auto-config functionality to also store default
values, while otherwise being compatible with the previous
implementation.

Additional context:

In the BayesFlow library, we provide models for downstream use, and aim
to avoid changes that break loading of models. We rely on the
auto-config functionality to avoid the somewhat tedious and error-prone
hand-written specification of the configuration. To be able to change
defaults without breaking model loading, it would be great if defaults
were serialized in the auto-config as well.

See also https://github.com/bayesflow-org/bayesflow/issues/566.

* auto-conf: correct behavior to match previous implementation

* auto-config: riase for variadic arguments, add tests

Variadic arguments (i.e., *args) and positional only arguments cannot be
passed via `cls(**config)`, so no automatic config can be created for
them without adding additional logic to `from_config`. This commit adds
an appropriate error message and tests for this case.

It also extends the test to contain an optional argument with a default
value. This allows testing the behavior defined in the previous commits.