Do not use `os.path.sep` to detect `/` in names.
Operation and layer names do not represent a filesystem path and should not be OS dependent. Namescope paths always use "/" as a separator, which is what we should be looking for.

This reverts https://github.com/keras-team/keras/pull/21343 just for `Operation`.

Note that [the unit test](https://github.com/keras-team/keras/blob/master/keras/src/ops/operation_test.py#L331) was correct. Simply, we never run in on Windows, which is why this was not caught.