Fix deadlock in `CallbackList`. (#21701)
A memory leak related to the executor in `CallbackList` was fixed in https://github.com/keras-team/keras/pull/20779

However, calling `Executor.shutdown` within `__del__` is intrisincally unsafe and can create deadlocks because the garbage collector can be called in different contexts.

This new approach uses the `on_train/test/predict_begin` and `on_train/test/predict_end` callbacks to detect when we're done with the executor.
- it counts the number of "begin"s and "end"s to handle the case of `evaluate` within `fit` (we do not shutdown the executor at the end of `evaluate` but instead keep it around for the rest of the training)
- it also handles `CallbackList` being reused between calls to `fit`, `evaluate` or `predict` even though Keras doesn't reuse.

Also renamed `_clear_futures` to `_flush_futures` to make it clear futures are not discarded, but exectuted.