Do not use backend ops in `ProgBar`. (#21709)
The `ProgBar` was using `backend.numpy.mean` causing it to use the accelerator (TPU or GPU) therefore causing a synchronization which defeated the async callback mechanism. This in turn was slowing down training.

All values provided in the `logs` are already Python floats and are all single values. There is therefore no need to do a `mean`, computing the average is simply a division of the running sum by the running count.