OpenVino `device_scope` and data adapters tests
The `device` scope feature was not working because of a missing import as well as incorrect device name logic. The existing device name logic had been copied from the torch implementation but is not applicable to OpenVino.

Also turned on all the data adapter related tests for OpenVino, which happen to exercise the `device` scope. Data adapters are used with `predict` and `evaluate` and are useful with OpenVino even without training support.