Always use `run_tpu_tests` label to run the TPU tests. (#21900)
Problem: the current approach creates two entries for TPU tests: one for `pull_request` and one for `pull_request_review`. But we only want one entry and one run of the tests.

Solution: add a new labeler workflow to label the PR on approval, which in turn triggers the TPU tests.

Also rename `welcome` workflow.