Remove nightly tests with Python 3.13. (#21943)
This is a revert of https://github.com/keras-team/keras/pull/21566

These tests have never passed: https://github.com/keras-team/keras/actions/runs/20325149006/job/58388669940

The issue is that some of our dependencies are not available for Python 3.13.

Additionally, nightly job are not monitored, so if they fail, it can go unnoticed for months.